add http status code

+++ src/main/java/com/jsp/lms/controller/BookController.java
@@ -2,7 +2,10 @@ package com.jsp.lms.controller;
 
 import com.jsp.lms.entiry.Book;
 import com.jsp.lms.service.BookService;
+import com.jsp.lms.utility.ResponseStructure;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,29 +20,89 @@ public class BookController {
     private BookService bookService;
 
    @PostMapping
-     public Book addBook(@RequestBody Book book){
-        return bookService.addBook(book);
+     public ResponseEntity<ResponseStructure<Book>> addBook(@RequestBody Book book){
+       book =  bookService.addBook(book);
+       ResponseStructure<Book> responseStructure = new ResponseStructure<>();
+       responseStructure.setStatusCode(HttpStatus.CREATED.value());
+       responseStructure.setMsg("Book Created");
+       responseStructure.setData(book);
+
+         return new ResponseEntity<ResponseStructure<Book>>(responseStructure,HttpStatus.CREATED);
     }
 
   @GetMapping
-    public List<Book> findAllBook(){
-        return bookService.findAllBook();
+    public ResponseEntity<ResponseStructure<List<Book>>> findAllBook(){
+        List<Book> books =  bookService.findAllBook();
+        ResponseStructure<List<Book>> responseStructure = new ResponseStructure<>();
+      if (books == null){
+          responseStructure.setStatusCode(HttpStatus.NOT_FOUND.value());
+          responseStructure.setMsg("Book not Found id:" );
+          responseStructure.setData(null);
+          return new ResponseEntity<>(responseStructure, HttpStatus.NOT_FOUND);
+      }
+        responseStructure.setStatusCode(HttpStatus.FOUND.value());
+      responseStructure.setMsg("Find All Book");
+      responseStructure.setData(books);
+      return new ResponseEntity<ResponseStructure<List<Book>>>(responseStructure,HttpStatus.FOUND);
    }
 
-  @GetMapping("/id")
-    public Book findById(@RequestParam int id){
-        return bookService.findById(id);
+  @GetMapping("/{id}")
+    public ResponseEntity<ResponseStructure<Book> > findById(@PathVariable int id){
+        Book book  =  bookService.findById(id);
+        ResponseStructure<Book> responseStructure = new ResponseStructure<>();
+      if (book == null) {
+          // Book not found
+          responseStructure.setStatusCode(HttpStatus.NOT_FOUND.value());
+          responseStructure.setMsg("Book not found with id: " + id);
+          responseStructure.setData(null);
+          return new ResponseEntity<>(responseStructure, HttpStatus.NOT_FOUND);
+      }
+
+      responseStructure.setStatusCode(HttpStatus.OK.value());
+        responseStructure.setMsg("Find Book by Id");
+        responseStructure.setData(book);
+
+        return new ResponseEntity<ResponseStructure<Book>>(responseStructure,HttpStatus.OK);
    }
 
 
-  @PutMapping("/id")
-    public Book updateById(@RequestParam int id, @RequestBody Book updateBook)  {
-        return bookService.updateById(id,updateBook);
+  @PutMapping("/{id}")
+    public ResponseEntity<ResponseStructure<Book>> updateById(@PathVariable int id, @RequestBody Book updateBook)  {
+       Book book = bookService.findById(id);
+      ResponseStructure<Book> responseStructure = new ResponseStructure<>();
+       if (book == null){
+           responseStructure.setStatusCode(HttpStatus.NOT_FOUND.value());
+           responseStructure.setMsg("Book not found with id: " + id);
+           responseStructure.setData(null);
+           return new ResponseEntity<>(responseStructure, HttpStatus.NOT_FOUND);
+        }else {
+           bookService.updateById(id,updateBook);
+           responseStructure.setStatusCode(HttpStatus.OK.value());
+           responseStructure.setMsg("Update by Id");
+           responseStructure.setData(book);
+           return  new ResponseEntity<>(responseStructure,HttpStatus.OK);
+       }
+
+
    }
 
- @DeleteMapping("/id")
-    public Book deleteById(@RequestParam int id){
-        return bookService.deleteById(id);
+ @DeleteMapping("/{id}")
+    public ResponseEntity<ResponseStructure<Book>>deleteById(@PathVariable int id){
+        Book book = bookService.findById(id);
+
+            ResponseStructure<Book> responseStructure = new ResponseStructure<>();
+            if (book == null ){
+                responseStructure.setStatusCode(HttpStatus.NOT_FOUND.value());
+                responseStructure.setMsg("Book not found with id: " + id);
+                responseStructure.setData(null);
+                return new ResponseEntity<>(responseStructure, HttpStatus.NOT_FOUND);
+            }else {
+                bookService.deleteById(id);
+                responseStructure.setStatusCode(HttpStatus.OK.value());
+                responseStructure.setMsg("Delete Book by Id");
+                responseStructure.setData(book);
+                return new ResponseEntity<>(responseStructure, HttpStatus.OK);
+            }
    }
 
 }
